RollingBloomDBGTest: change remaining size_t to uint64_t

+++ Unittest/BloomDBG/RollingBloomDBGTest.cpp
@@ -156,7 +156,7 @@ TEST_F(RollingBloomDBGTest, pathTraversal)
 	const V GACTC("GACTC", RollingHash("GACTC", m_numHashes, m_k));
 	const V ACTCG("ACTCG", RollingHash("ACTCG", m_numHashes, m_k));
 
-	size_t hashes[MAX_HASHES];
+	uint64_t hashes[MAX_HASHES];
 	CGACT.rollingHash().getHashes(hashes);
 	bloom.insert(hashes);
 	GACTC.rollingHash().getHashes(hashes);
@@ -268,7 +268,7 @@ protected:
 		 * any additional edges in the graph.
 		 */
 
-		size_t hashes[MAX_HASHES];
+		uint64_t hashes[MAX_HASHES];
 		RollingHash("CGACT", m_numHashes, m_k).getHashes(hashes);
 		m_bloom.insert(hashes);
 		RollingHash("TGACT", m_numHashes, m_k).getHashes(hashes);
